==> 0_SQLconversion.ipynb <==
import sqlite3
import pandas as pd
import matplotlib.pyplot as plt
import numpy as np

from sklearn import tree, preprocessing
import sklearn.ensemble as ske
from sklearn.model_selection import train_test_split

# Input data files are available in the "../input/" directory.
# For example, running this (by clicking run or pressing Shift+Enter) will list the files in the input directory

from subprocess import check_output
#print(check_output(["ls", "../input"]).decode("utf8"))

# Any results you write to the current directory are saved as output.

cnx = sqlite3.connect('./FPA_FOD_20170508.sqlite')

import os.path

df = pd.read_sql(
                       """
                       
                        SELECT *
                        from fires 
                        
                       """, con=cnx)

columns = ['DISCOVERY_DOY', 'DISCOVERY_DATE', 'FIRE_YEAR', 'DISCOVERY_TIME', 'STAT_CAUSE_CODE', 'CONT_DATE', 'CONT_DOY', 'FIRE_SIZE_CLASS', 'FIRE_SIZE', 'LATITUDE', 'LONGITUDE', 'STATE']
df=df[columns]
df.to_csv('sql.csv')
