==> 2_Data_Preparation_Cleaning.ipynb <==
# Data Frame Work

### Load Pandas and Get DataFrame Setup Properly

### Please note I was not able to upload original csv because of its size to github. go to https://www.kaggle.com/rtatman/188-million-us-wildfires if you want to get it from a in SQL. I converted it from SQL to csv and then refined it down to a more managable size which you will see throughout the report.

#import pandas, all that is needed for this file
import pandas as pd

#load csv as df
df = pd.read_csv('sql.csv')

#### Round data to nearest 5, I want usa to be categorized like a checker board by 5 latitude by 5 longitude squares. Latitude is ~70 miles and Longitude is ~ 55 miles. Then concat new values with data.

df.head()

#chose to append rather than df2['lat'][i] = i because of how much faster it is.
latty5 = []
latty2 = []
latty1 = []
longy5 = []
longy2 = []
longy1 = []

for y, i in enumerate(df['LATITUDE']):
    if y%2500==0:
        print(y)
    latty5.append(int(5*round(float(i)/5)))
    latty2.append(int(2*round(float(i)/2)))
    latty1.append(int(1*round(float(i)/1)))
    longy1.append(int(1*round(float(df['LONGITUDE'][y])/1)))
    longy2.append(int(2*round(float(df['LONGITUDE'][y])/2)))
    longy5.append(int(5*round(float(df['LONGITUDE'][y])/5)))

#convert rounded lattitudes to dataframe
df5=pd.DataFrame(latty5).rename(columns={0: 'latt5'})
df6 = pd.DataFrame(latty2).rename(columns={0: 'latt2'})
df7= pd.DataFrame(latty1).rename(columns={0: 'latt1'})
df8=pd.DataFrame(longy1).rename(columns={0: 'longg1'})
df9=pd.DataFrame(longy2).rename(columns={0: 'longg2'})
df11=pd.DataFrame(longy5).rename(columns={0: 'longg5'})

#changed over to df10, wanted to keep the df and df2 just incase something needs to be changed I did not want to overwrite it
df10 = pd.concat([df5, df], axis=1)
df10=pd.concat([df6, df10], axis=1)
df10=pd.concat([df7, df10], axis=1)
df10=pd.concat([df8, df10], axis=1)
df10=pd.concat([df9, df10], axis=1)
df10=pd.concat([df11, df10], axis=1)

# Filter data so I can condense the latitudes and longitudes into rectangle and important/ relevant details
#### Reason filtering as method is that a for loop would be too slow. Instead by filtering by date, latitude and longitude I can go through if statements that are far quicker and more efficient. (20 Minutes... Not 15 hours)

df10.head()

#sort data by discovery date, then latitude, then longitude.
df10=df10.sort_values(['DISCOVERY_DATE', 'latt2', 'longg2'])

#cut out Alaska. Felt that I was concerned about the continental USA and highly populated regions. 
#Getting patterns for USA is more significant!

df10 = df10[df10['LATITUDE']<50]
df10=df10[df10['LONGITUDE']>=-150]

#reset index! very important in these cases
df10=df10.reset_index().drop(columns=['index'])

df20 = df10

df10.head()

##data is ordered by date, latitude and longitude. through if statements filtering the data into individual collumns.

#different series that will become dataframes
series2dates=[]
f_lat = []
f_lon = []
series6sz_cl = []
series7_state = []
series8_cause = []
series9_firesize=[]
gf_latty = []
gf_longy = []


for i,y in enumerate(df10['latt2']):
    if i%25000 == 0:
        print(i)
        #print at 25000 because I do not want to use up too much memory by constantly printing i, but I do want to keep track
        #because sometimes something breaks and it is hard to tell/ it is hard to tell how long it will take time wise
    if i == 0:      
        #i==0 is different than the rest of the data, and everything is based the data that is ==0.
        #creating lists of size class, state, longitude, latitude, cause, fire size
        #plugging in the rounded latitudes and longitudes and the date, for these there should not be multiple 
        sz_cl = []
        sz_cl=str(df10['FIRE_SIZE_CLASS'][i])
        
        state_data = []
        state_data=str(df10['STATE'][i])
        
        df_longy = []
        df_longy=str(df10['LONGITUDE'][i])
        
        df_latty = []
        df_latty=str(df10['LATITUDE'][i])

        df_cause = []
        df_cause = str(df10['STAT_CAUSE_CODE'][i])
        
        df_fire = []
        df_fire = str(df10['FIRE_SIZE'][i])
        

        f_lat.append(df10['latt2'][i])        
        f_lon.append(df10['longg2'][i])
        series2dates.append(df10['DISCOVERY_DATE'][i])
               
        
    elif df10['DISCOVERY_DATE'][i] == df10['DISCOVERY_DATE'][(i-1)]:
        if df10['latt2'][i] == df10['latt2'][i-1]:
            if df10['longg2'][i] == df10['longg2'][i-1]:
                ##these ifs mean that that if the date, latitude and longitude from one row match the row prior, this means
                #that there was no change, so either it was to do nothing or to add it to a string (depending on the data)
                sz_cl=str(df10['FIRE_SIZE_CLASS'][i])+','+str(sz_cl)

                df_latty=str(df10['LATITUDE'][i])+','+str(df_latty)
                df_longy = str(df10['LONGITUDE'][i])+','+str(df_longy)
                
                state_data=str(df10['STATE'][i])+','+str(state_data)

                df_cause=str(df10['STAT_CAUSE_CODE'][i])+','+str(df_cause)

                df_fire = str(df_fire)+','+str(df10['FIRE_SIZE'][i])
                
                #adding to lists, rounded latitude/longitude and date are left out on purpose. only one needed.
                #everything else is included. this includes fire size class, latitude, longitude, state, cause, and fire size
                #using strings because append was turning every new point added to a new column, when I wanted it 
                #to be within a string
                

            else:
                ##all elses are the same, essentially there is some sort of change which means that it is completely different
                #so all the data needs to be compltelly updated. this is where we append these data to the original dataframe
                #these are all in the order i want. 
                
                #i also am trying to refresh the data, with the Xyz = [] and then Xyz=str(df10['Variable'][i]) it is doing that
                #all elses are the same, so no need to repeat
                series2dates.append(df10['DISCOVERY_DATE'][i])

                f_lat.append(df10['latt2'][i])
                f_lon.append(df10['longg2'][i])

                gf_latty.append(df_latty)
                df_latty = []
                df_latty=str(df10['LATITUDE'][i])
        
                gf_longy.append(df_longy)
                df_longy = []
                df_longy=str(df10['LONGITUDE'][i])
        
        
                series7_state.append(state_data)
                state_data = []
                state_data=str(df10['STATE'][i])

                series8_cause.append(df_cause)
                df_cause = []
                df_cause=str(df10['STAT_CAUSE_CODE'][i])+','+str(df_cause)

                series6sz_cl.append(sz_cl)
                sz_cl = []
                sz_cl=str(df10['FIRE_SIZE_CLASS'][i])

                series9_firesize.append(df_fire)
                df_fire = []
                df_fire=str(df10['FIRE_SIZE'][i])


#all elses are the same, so no need to repeat. listed above.
        
        else:
            
            series2dates.append(df10['DISCOVERY_DATE'][i])
            
            f_lat.append(df10['latt2'][i])
            f_lon.append(df10['longg2'][i])

            gf_latty.append(df_latty)
            df_latty = []
            df_latty=str(df10['LATITUDE'][i])

            gf_longy.append(df_longy)
            df_longy = []
            df_longy=str(df10['LONGITUDE'][i])

        

                                  
            series7_state.append(state_data)
            state_data = []
            state_data=str(df10['STATE'][i])
            
            series8_cause.append(df_cause)
            df_cause = []
            df_cause=str(df10['STAT_CAUSE_CODE'][i])+','+str(df_cause)

            series6sz_cl.append(sz_cl)
            sz_cl = []
            sz_cl=str(df10['FIRE_SIZE_CLASS'][i])
            
            series9_firesize.append(df_fire)
            df_fire = []
            df_fire=str(df10['FIRE_SIZE'][i])
            
            
#all elses are the same, so no need to repeat. listed above.
            
    else:
                
        series2dates.append(df10['DISCOVERY_DATE'][i])
        
        
        f_lon.append(df10['longg2'][i])
        f_lat.append(df10['latt2'][i])
        
        gf_latty.append(df_latty)
        df_latty = []
        df_latty=str(df10['LATITUDE'][i])

        gf_longy.append(df_longy)
        df_longy = []
        df_longy=str(df10['LONGITUDE'][i])

        
            
        
        series8_cause.append(df_cause)
        df_cause = []
        df_cause=str(df10['STAT_CAUSE_CODE'][i])+','+str(df_cause)
        
        
        series6sz_cl.append(sz_cl)
        sz_cl = []
        sz_cl=str(df10['FIRE_SIZE_CLASS'][i])
            
        series7_state.append(state_data)
        state_data = []
        state_data=str(df10['STATE'][i])

        
        series9_firesize.append(df_fire)
        df_fire = []
        df_fire=str(df10['FIRE_SIZE'][i])
            
         
        
        

#turn all the different datapoints as dataframes. 
dates = pd.DataFrame(series2dates).rename(columns={0: 'dates'})
latit = pd.DataFrame(f_lat).rename(columns={0: 'latit'})
longit = pd.DataFrame(f_lon).rename(columns={0: 'longit'})
latty2 = pd.DataFrame(gf_latty).rename(columns={0: 'latitude'})
longy2 = pd.DataFrame(gf_longy).rename(columns={0: 'longitude'})
sz_class = pd.DataFrame(series6sz_cl).rename(columns={0: 'sz_class'})
state = pd.DataFrame(series7_state).rename(columns={0: 'state'})
cause = pd.DataFrame(series8_cause).rename(columns={0: 'cause'})
firesize = pd.DataFrame(series9_firesize).rename(columns={0: 'firesize'})
latty2 = pd.DataFrame(gf_latty).rename(columns={0: 'latitude'})
longy2 = pd.DataFrame(gf_longy).rename(columns={0: 'longitude'})

#turn all dataframes into one big dataframe. can see by the shape that it changed majorly
df1000 =  pd.concat([dates, latit], axis=1)
df1000 =  pd.concat([df1000, longit], axis=1)
df1000 =  pd.concat([df1000, longy2], axis=1)
df1000 =  pd.concat([df1000, latty2], axis=1)
print('stop')
df1000 =  pd.concat([df1000, state], axis=1)
df1000 =  pd.concat([df1000, sz_class], axis=1)
df1000 =  pd.concat([df1000, cause], axis=1)
df1000 =  pd.concat([df1000, firesize], axis=1)

#save to csv. changing over to a new book because of memory issues
df1000.to_csv('GoodDataFrame.csv')
