==> 3_Data_Preparation_Clean.ipynb <==
import pandas as pd

df1000 = pd.read_csv('GoodDataFrame.csv')

# Create DataFrames that will be concated with Final DataFrame
###### A,B,C...G are all different classifications of fire sizes, I summed each appearance of a letter grade. If I count 10 A's on a certain day at a certain location, then it ='s 10.

#getting dataframe started. if there is no observation then it =='s zero thus 0 is the optimal choice
df1000['A'] = 0
df1000['B'] = 0
df1000['C'] = 0
df1000['D'] = 0
df1000['E'] = 0
df1000['F'] = 0
df1000['G'] = 0

#import counter
from collections import Counter

#build A,B...G count. Need Key, Value pairs to do this. Use counter and .items() to access that information
#think of this like dummy variables. the function when it is finished breaks rather than ends smoothly but still very happy with it
for i,y in enumerate(df1000['sz_class']):
    if i%100000==0:
        print(i)
    for key, value in dict(Counter(df1000['sz_class'][i])).items(): 
        if key == 'A':
            df1000['A'][i] = value
        if key == 'B':
            df1000['B'][i] = value
        if key == 'C':
            df1000['C'][i] = value
        if key == 'D':
            df1000['D'][i] = value
        if key == 'E':
            df1000['E'][i] = value
        if key == 'F':
            df1000['F'][i] = value
        if key == 'G':
            df1000['G'][i] = value

df1000.to_csv('df1010.csv')

pd.read_csv('ABC_tuenight.csv').head()

# Create a Max Value Column and Sum of Values Column
#### Each row needs a max value, and sum of all the wildfiries. This is important details that will be leveraged later for analysis

#function for finding max
#just like the function prior this function breaks
append1 = []
append2 = []

for i,y in enumerate(df1000['firesize']):
    if i%100000==0:
        #print i if divisable by 1000 because not want to use too much memory but want to track work
        print(i)
    #add append1 to append2 and then clear append1, and t, but not append2
    append2.append(append1)
    append1 = []
    t = 0
    ### separate the different units in the firesize by splitting it. then use a function to find the max values and 
    ##continously cycle through. 
    for m in df1000['firesize'].tolist()[i].split(","):
            if float(m) > float(t):
                t = float(m)
                append1 = []
                append1.append(t)
#             else:
#                 print('hi')
# #         else:
#             if float(m) > float(t):
#                 append1 = []
#                 append1.append(float(m))
                            

#create 2 of the dataframes, add them lower to the Final dataframe
dataframe = pd.DataFrame(append2)
firessmax = pd.DataFrame(append2).rename(columns={0: 'FIRE SIZE MAX'})

#sum the acreage of the fires
append10 = []
append20 = []

for i,y in enumerate(df1000['firesize']):
    if i%200000==0:
        print(i)
    #add when change of i, add prior append10 to append 20
    #clear t and append10
    append20.append(append10)
    append10 = []
    t = 0
    
    #use .tolist()[i].split(','), clear append10, which gets each float alone, then add them together and then add it to append10
    for m in df1000['firesize'].tolist()[i].split(","):
        append10 = []
        t = float(t)+float(m)
        append10.append(t)
        


#convert them to csv out of paranoia and create dataframes and concat it with the final dataframe
firesssum = pd.DataFrame(append20).rename(columns={0: 'FIRE SIZE SUM'})

#reset the indexes it can be easy to run into trouble if you forget to do that. I also dropped the unncessary column that comes
#when you change the index
firessum = firesssum.drop(0).reset_index().drop(columns='index')

#### the last and first dataframe were missing, so I added those in manually for all the various examples
#### same applies to fire size max see below

firessum['FIRE SIZE SUM'][0] = 3.0

firessum['FIRE SIZE SUM'][544534] = 0

df1000['FIRE SIZE SUM'] = firessum['FIRE SIZE SUM']

firessmax = firessmax.drop(0).reset_index().drop(columns='index')

firessmax['FIRE SIZE MAX'][0] = 3.0

firessmax['FIRE SIZE MAX'][544534] = 0

df1000['FIRE SIZE MAX'] = firessmax['FIRE SIZE MAX']

justincase=pd.concat([firessmax, firesssum])

import numpy as np

##### Using .diff to see difference between current value and values prior. This is SUPER IMPORTANT. These will be my y variables. I am not interested in finding when will there be a lot of forest fire but rather when will there be a change in conditions that will cause a change and expansion of forest fires.

##### I then took only the max value of these columns so that it is standardized nicely and consistant. 

#### arguably the most important function of the whole project. This is me creating the y variable. I am saying that if there is no fire and it jumps to 100 acres, or the sum of fire in an area grows by 4x or grows by 2x under certain conditions then to make note of it and make them 1's. Everything else is a 0. 

#### y variable is named firebinary.

df1000.head()

df1000.columns

df100=pd.read_csv('datesandstuffandweekdays.csv')

df100.rename(columns={'Unnamed: 3': 'Julian'}, inplace=True)

#### this is to lable the day of the week. It could be a significant detail, so I am getting monday, tue, and whether they are a weekday or not, and what day it is by normal year standards. The details originally were in Julian calendars and were a bit uninterpretable. 

dayoweek = []
dato = []
weekend = []
for i,y in enumerate(df1000['dates']):
    for m,q in enumerate(df100['Julian']):
        if q == y:

            dayoweek.append(df100['date'][m])
            dato.append(df100['day of week'][m])
            if str(df100['day of week'][m]) == 'Friday' or str(df100['day of week'][m]) == 'Tuesday' or str(df100['day of week'][m]) == 'Wednesday' or str(df100['day of week'][m]) == 'Thursday':
                weekend.append(0)
                if i%150000==0:
                    print(i)
                    print(df100['date'][m])
                    print(df100['day of week'][m])
            else: 
                weekend.append(1)
                if i%150000==0:
                    print(i)
                    print(df100['date'][m])
                    print(df100['day of week'][m])
                    
            

dater3 = pd.DataFrame(weekend).rename(columns={0: 'weekend'})

dater2 = pd.DataFrame(dato).rename(columns={0: 'normal date'})

dater1 = pd.DataFrame(dayoweek).rename(columns={0: 'day of week'})

df1000=pd.concat([df1000, dater3], axis=1)
df1000=pd.concat([df1000, dater1], axis=1)
df1000=pd.concat([df1000, dater2], axis=1)

df1000=df1000.drop(columns=['Unnamed: 0'])

##### create a name for each individual category. I did this by combining each the cordinates into this name. Each one is unique to the different potential coordinates. This helps in categorizing and organizing. 

##latitude and longitude appended. 
lattlon=[]
for i,y in enumerate(df1000['latit']):
    ###lattlon adds the the specific values corresponding so they can be dividided up into vairous categories
    lattlon.append('lat:'+str(y)+'lon:'+str(df1000['longit'][i]))
    if i%150000==0:
            print('lat:'+str(y)+'lon:'+str(df1000['longit'][i]))
            print(i)

lat_lon = pd.DataFrame(lattlon).rename(columns={0: 'lat lon'})

#### Some format got a little messed up, so I corrected those details and incorporarted the latlon titles into the main dataframe

df1000=pd.concat([df1000, lat_lon], axis=1)

df1000=df1000.sort_values('latit')

df1000.reset_index(inplace=True)

df1000.drop(544534, inplace=True)

df1000.rename(columns={'normal date':'day of week', 'day of week': 'day'}, inplace=True)

df1000['day']=pd.to_datetime(df1000['day'])

df10000=df1000.sort_values(['lat lon', 'dates'])

df10000.to_csv('datacleaned.csv')
