==> 4_data_cleaning.ipynb <==
import pandas as pd

df1000=pd.read_csv('datacleaned.csv')

df1000.drop(columns='diffmax', inplace=True)

### re prep data

##### clear out all data below 2003. Looking at it it was far more sparingly collected and less consistent. I did not feel it was logical to assume that there are just so much more wildfires occuring today, but rather the ability to record or something has dramatically changed, thus I decided to remove it. 

#convert df1000 to datetime
df1000['day']=pd.to_datetime(df1000['day'])

#change dataframe to be above a certain date. it became apparent that data before 2003 was far less standardized than the
#data after
df500 = df1000[df1000['dates']>2452640.5]

df500.reset_index(inplace=True)

df500.drop(columns=['Unnamed: 0', 'index', 'level_0', 'Unnamed: 0.1', 'Unnamed: 0.1.1'], inplace=True)

#convert data as type int
df500['dates']=df500['dates'].astype(int)

# need to do range with floats or figure out why they julian days are repeating

# Reset Index and Fill in Missing Columns
#### Need to look at historical data and put it in rows. In order to do it efficiently and accurately I need to inclue all dates even the ones that are blanks, and I need to have the data ordered by 'lat lon' and then by date. through that I can use if then statements to filter data how I want which you will see further down


#### Below is me working with resetting the index and getting all the data points in properly included. So even though I did not have an observation on a given day, it is still in the dataframe and in order and accounted for. Julian Calendars made this far easier. 

#this was a doosy
#first looping through value counts on a reset index so that I can pull out and loop through all the latitude and longitudes 
#and automatically plug them in

#second loop I reindexed the data properly on the dates (julian calendar)

for i,y in enumerate(df500['lat lon'].value_counts().reset_index()['index']):
    idx = range(2452641, 2457387, 1)
    if i == 0:
        df5000 = df500.set_index('dates')
        s = 0
        s = df5000[df5000['lat lon']==y]
        z = 0
        z = s.reindex(idx, fill_value = 0)
    else:
        df2500 = df500.set_index('dates')
        s = 0
        s = df2500[df2500['lat lon']== y]
        q = 0
        q = s.reindex(idx, fill_value = 0)
        z=pd.concat([z,q], axis=0)

#after all that beautiful work is done, definitely want to reset the index
z.reset_index(inplace=True)

z['A_']=z['A'].shift(-1)
z['B_']=z['B'].shift(-1)
z['C_']=z['C'].shift(-1)
z['D_']=z['D'].shift(-1)
z['E_']=z['E'].shift(-1)
z['F_']=z['F'].shift(-1)
z['G_']=z['G'].shift(-1)

z['FIRE SIZE SUM_']=z['FIRE SIZE SUM'].shift(-1)

z['FIRE SIZE MAX_']=z['FIRE SIZE MAX'].shift(-1)

z.columns

z['A_45']=z['A'].rolling(window=30).sum().shift(-60)
z['B_45']=z['B'].rolling(window=30).sum().shift(-60)
z['C_45']=z['C'].rolling(window=30).sum().shift(-60)
z['D_45']=z['D'].rolling(window=30).sum().shift(-60)
z['E_45']=z['E'].rolling(window=30).sum().shift(-60)
z['F_45']=z['F'].rolling(window=30).sum().shift(-60)
z['G_45']=z['G'].rolling(window=30).sum().shift(-60)
z['FIRE SIZE SUM_45']=z['FIRE SIZE SUM'].rolling(window=30).sum().shift(-60)
z['FIRE SIZE MAX_45']=z['FIRE SIZE MAX'].rolling(window=30).sum().shift(-60)

z['A_365']=z['A'].rolling(window=90).sum().shift(-400)
z['B_365']=z['B'].rolling(window=90).sum().shift(-400)
z['C_365']=z['C'].rolling(window=90).sum().shift(-400)
z['D_365']=z['D'].rolling(window=90).sum().shift(-400)
z['E_365']=z['E'].rolling(window=90).sum().shift(-400)
z['F_365']=z['F'].rolling(window=90).sum().shift(-400)
z['G_365']=z['G'].rolling(window=90).sum().shift(-400)
z['FIRE SIZE SUM_365']=z['FIRE SIZE SUM'].rolling(window=90).sum().shift(-400)
z['FIRE SIZE MAX_365']=z['FIRE SIZE MAX'].rolling(window=90).sum().shift(-400)

z['A_2yr']=z['A'].rolling(window=365).sum().shift(-912)
z['B_2yr']=z['B'].rolling(window=365).sum().shift(-912)
z['C_2yr']=z['C'].rolling(window=365).sum().shift(-912)
z['D_2yr']=z['D'].rolling(window=365).sum().shift(-912)
z['E_2yr']=z['E'].rolling(window=365).sum().shift(-912)
z['F_2yr']=z['F'].rolling(window=365).sum().shift(-912)
z['G_2yr']=z['G'].rolling(window=365).sum().shift(-912)
z['FIRE SIZE SUM_2yr']=z['FIRE SIZE SUM'].rolling(window=90).sum().shift(-912)
z['FIRE SIZE MAX_2yr']=z['FIRE SIZE MAX'].rolling(window=90).sum().shift(-912)

z['A_7']=z['A'].rolling(window=7).sum().shift(-7)
z['B_7']=z['B'].rolling(window=7).sum().shift(-7)
z['C_7']=z['C'].rolling(window=7).sum().shift(-7)
z['D_7']=z['D'].rolling(window=7).sum().shift(-7)
z['E_7']=z['E'].rolling(window=7).sum().shift(-7)
z['F_7']=z['F'].rolling(window=7).sum().shift(-7)
z['G_7']=z['G'].rolling(window=7).sum().shift(-7)
z['FIRE SIZE SUM_7']=z['FIRE SIZE SUM'].rolling(window=7).sum().shift(-7)
z['FIRE SIZE MAX_7']=z['FIRE SIZE MAX'].rolling(window=7).sum().shift(-7)

# Sum of Max Fire Acreage and Sum of total fire Acreage

#### did this over various intervals. turns out the time interval can be too close and not really helpful when you are within 10 days.
#### did incraments of 4 years before, 1 year before, and 45 days before. These are obviously each a range but the data was extremely helpful. ALSO this means that I will have to cut 4 years later on because with a 4 year lag there is not data on those. 

##### Using .diff to see difference between current value and values prior. This is SUPER IMPORTANT. These will be my y variables. I am not interested in finding when will there be a lot of forest fire but rather when will there be a change in conditions that will cause a change and expansion of forest fires.

##### I then took only the max value of these columns so that it is standardized nicely and consistant. 

diff2 = pd.DataFrame(z['FIRE SIZE SUM'].diff(2))
diff3=pd.DataFrame(z['FIRE SIZE SUM'].diff(3))
diff5=pd.DataFrame(z['FIRE SIZE SUM'].diff(5))
diff4=pd.DataFrame(z['FIRE SIZE SUM'].diff(4))
diff1=pd.DataFrame(z['FIRE SIZE SUM'].diff(1))
diff6=pd.DataFrame(z['FIRE SIZE SUM'].diff(6))

diff=pd.concat([diff2, diff3], axis=1)
diff=pd.concat([diff, diff4], axis=1)
diff=pd.concat([diff, diff5], axis=1)
diff=pd.concat([diff, diff1], axis=1)
diff=pd.concat([diff, diff6], axis=1)

diffmax=diff.max(axis=1)
diffmax=diffmax.rename(columns={0: 'diffmax'})

z=pd.concat([z, diffmax], axis=1)
z.rename(columns={0: 'diffmax'}, inplace=True)

#save to csv
z.to_csv('lagging.csv')

#the way it autofilled earlier, '1970-01-01' was the documentations way of saying na. I cut out all values of this kind
#also removed values beforew 2007. reason for that is they do not have the lagged information and details so the variables
#would have been all messed up.

#this also accounts for me making this shift accurate as all of the details will align as the dataframe was uniform and aligned
#by location. so the dates lined up perfectly. more efficient than a forloop

##2454101== Dec 31, 2006. 
z = z[z['day']!='1970-01-01']
dfzqq = z
z=z[z['dates']>2453371]
z.reset_index(inplace=True)

#convert day to datetime
z['day']=pd.to_datetime(z['day'])

dfzqq['day'] = pd.to_datetime(z['day'])

#pull out the year month and day of month so that they can be represented properly in columns
year2 = []
month2 = []
dayom2 = []
for i,y in enumerate(dfzqq['day']):
    year2.append(str(y)[:4])
    month2.append(str(y)[5:7])
    dayom2.append(str(y)[8:10])

#pull out the year month and day of month so that they can be represented properly in columns
year = []
month = []
dayom = []
for i,y in enumerate(z['day']):
    year.append(str(y)[:4])
    month.append(str(y)[5:7])
    dayom.append(str(y)[8:10])

#convert them into dataframes, ints and name them
year = pd.DataFrame(year).rename(columns={0: 'year'}).astype(int)
month = pd.DataFrame(month).rename(columns={0:'month'}).astype(int)
dayom = pd.DataFrame(dayom).rename(columns={0: 'day_of_month'}).astype(int)

#convert them into dataframes, ints and name them
year2 = pd.DataFrame(year2).rename(columns={0: 'year'})
month2 = pd.DataFrame(month2).rename(columns={0:'month'})
dayom2 = pd.DataFrame(dayom2).rename(columns={0: 'day_of_month'})

df3500=pd.concat([dfzqq, year2], axis=1)
df3500 = pd.concat([df3500, month2], axis=1)
df3500 = pd.concat([df3500, dayom2], axis=1)

df3500.fillna(0, inplace=True)

#put all together into a csv
df3000=pd.concat([z, year], axis=1)
df3000=pd.concat([df3000, month], axis=1)
df3000=pd.concat([df3000, dayom], axis=1)

df3000.to_csv('lagging.csv')
