==> 5__modeling.ipynb <==
import pandas as pd

#### leverage logistic regression as this is classification

#load in data
df = pd.read_csv('lagging.csv')

#import classification, traintestsplit, and counter
from collections import Counter
from sklearn.datasets import make_classification
import numpy as np
from sklearn.model_selection import train_test_split

## Create Dynamic Classifications Through Scoring System as to what is considred a large jump

firesizecriteria=[]
for i,y in enumerate(df['FIRE SIZE MAX']):

    if df['diffmax'][i] > 20000:
            firesizecriteria.append(1)
    elif df['diffmax'][i]*5 < df['FIRE SIZE MAX'][i]:
        if df['diffmax'][i] > 600:
            firesizecriteria.append(1)
        else:
            firesizecriteria.append(0)
    elif df['diffmax'][i]*2 < float(df['FIRE SIZE MAX'][i]):
        if df['diffmax'][i] > 1000:
            firesizecriteria.append(1)
        else:
            firesizecriteria.append(0)

            
    elif df['diffmax'][i]*4 < float(df['FIRE SIZE MAX'][i]):
        if df['diffmax'][i] > 500:
            firesizecriteria.append(1)
        else:
            firesizecriteria.append(0)
    elif df['diffmax'][i]*100 < float(df['FIRE SIZE MAX'][i]):
        if df['diffmax'][i] > 200:
            firesizecriteria.append(1)
        else:
            firesizecriteria.append(0)
    
    else:
        firesizecriteria.append(0)

#     else:
#         print(0)   
#         firesizecriteria.append(0)

### Drop Unncessary Columns

df=pd.concat([df, pd.DataFrame(firesizecriteria).rename(columns={0:'firesize_criteria'})], axis=1)

df5 = pd.get_dummies(df['day of week'], dummy_na=0)

df=pd.concat([df,df5], axis=1)

df55=df.dropna()

df55.columns

df55=df55.sort_values('dates')

df55=df55.reset_index().drop(columns='level_0')

# Estabolish a y variable

y = df55['firesize_criteria']

df55=df55.drop(columns=['lat lon', 'Unnamed: 0', 'Unnamed: 0.1', 'Unnamed: 0.1.1', 'Unnamed: 0.1.1.1', 'index', 'Unnamed: 0.1.1.1.1'])

df55=df55.drop(columns=['day', 'firesize_criteria'])

df55=df55.drop(columns=['day of week', 'cause', 'sz_class', 'state', 'longitude', 'latitude', 'firesize', 'df_30_A_60_sum.1', 'df_30_B_60_sum.1', 'df_30_C_60_sum.1', 'df_30_D_60_sum.1', 'df_30_E_60_sum.1', 'df_30_F_60_sum.1', 'df_30_G_60_sum.1', 'df_30_60_sum.1', 'df_30_60_max.1'])

# Create X and Y variables split on Date. 

### This is to train based on dates. Test is assumed to be in the future.

X_train=df55[:int(df55.shape[0]*.7)]
X_test=df55[int(df55.shape[0]*.7):]

##### Standardize The Data

from sklearn.preprocessing import StandardScaler
scaler = StandardScaler()
scaler.fit(X_train)
X_train=scaler.transform(X_train)
X_test=scaler.transform(X_test)

y_train=y[:int(df55.shape[0]*.7)]
y_test=y[int(df55.shape[0]*.7):]

y_train.shape

df55.shape

# Make classifications of Data

X_train, y_train = make_classification(n_classes=2, class_sep=2, weights=[.7,.3], n_samples=1700 )

X_test, y_test = make_classification(n_classes=2, class_sep=2, weights=[.7,.3], n_samples=300 )

# Run Keras to Use Neural Network that Classifies

from imblearn.under_sampling import RandomUnderSampler # doctest: +NORMALIZE_WHITESPACE
rus = RandomUnderSampler(ratio = .5)
rus.fit(X_train, y_train)

from imblearn.under_sampling import RandomUnderSampler # doctest: +NORMALIZE_WHITESPACE
rus = RandomUnderSampler(ratio = .5)
rus.fit(X_test, y_test)

from keras.utils.np_utils import to_categorical

from keras.models import Sequential
from keras.layers import Dense

model = Sequential()

from keras.utils.np_utils import to_categorical
y_train = to_categorical(y_train)
y_test = to_categorical(y_test)

###put in input, outputs and hidden sizes
n_input = X_train.shape[1]
n_hidden = n_input
n_output = y_train.shape[1]

from keras.layers import Dropout

#add key features
model.add(Dropout(0.2))
model.add(Dense(n_hidden, input_dim=500, activation='relu'))
model.add(Dropout(0.5))
model.add(Dense(n_hidden, input_dim=100, activation='relu'))
model.add(Dropout(0.5))
model.add(Dense(n_hidden, input_dim=1000, activation='relu'))
model.add(Dense(n_output, activation='sigmoid'))

#design model
model.compile(loss='binary_crossentropy', optimizer='adam', metrics=['accuracy'])

#enter in keras model
history = model.fit(X_train, y_train, validation_data=(X_test, y_test),
                    epochs=25, batch_size=None)#, batch_size=None, verbose=0)


from sklearn.metrics import classification_report, confusion_matrix

confusion_matrix(np.argmax(y_test, axis=1), model.predict_classes(X_test))
